move previews code into deploy-helpers
