+++ turbo.json
@@ -26,7 +26,10 @@
 		"WRANGLER_LOG",
 		"CLOUDFLARE_API_TOKEN",
 		"CLOUDFLARE_ACCOUNT_ID",
-		"WRANGLER_WRITE_LOGS"
+		"WRANGLER_WRITE_LOGS",
+		"GITHUB_HEAD_REF",
+		"GITHUB_REF_NAME",
+		"CI_COMMIT_REF_NAME"
 	],
 	"tasks": {
 		"dev": {
